A-CodeQuality
Methods still supports unused text-base Ui
JavaDocs for many methods was barebones.
Code readibility can be further improved.

Deprecate text-based Ui functionalities and dependencies
Add JavaDocs to all public functions and non trivial private methods.
Refractor code to improve readibility

Refractoring code improves readibility and clarity of code thus allows
programme to be understood by developers working on it.
