+++ config/checkstyle/checkstyle.xml
@@ -37,7 +37,7 @@
     <!-- All Java AST specific tests live under TreeWalker module. -->
     <module name="TreeWalker">
 
-        <!-- Required to allow exceptions in code style -->
+        <!-- Required to allow duke.exceptions in code style -->
         <module name="SuppressionCommentFilter">
             <property name="offCommentFormat" value="CHECKSTYLE.OFF\: ([\w\|]+)"/>
             <property name="onCommentFormat" value="CHECKSTYLE.ON\: ([\w\|]+)"/>
