+++ data/duke.txt
@@ -1,4 +1,4 @@
 E|0|project meeting|2019-10-15|2019-10-15
 D|0|return book|2019-10-15
-E|0|project meeting|2019-10-15|2019-10-15
+E|1|project meeting|2019-10-15|2019-10-15
 T|0|hehee
