+++ src/main/java/duke/commands/ExitCommand.java
@@ -1,0 +1,13 @@
+package duke.commands;
+
+import duke.tasks.TaskList;
+import duke.utils.Storage;
+
+public class ExitCommand extends Command {
+    private static final String EXIT_MSG = "Goodbye Courier!";
+
+    @Override
+    public String execute(TaskList tasks, Storage storage) {
+        return EXIT_MSG;
+    }
+}
