+++ src/main/java/duke/exceptions/InvalidInputException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package duke.exceptions;
 
 public class InvalidInputException extends DukeException {
