+++ src/main/java/javafx_utils/DialogBox.java
@@ -1,3 +1,5 @@
+package javafx_utils;
+
 import java.io.IOException;
 import java.util.Collections;
